建造：显示

+++ AdmiralRoom/API/getmember_kdock.cs
@@ -5,7 +5,7 @@
         public int api_id { get; set; }
         public int api_state { get; set; }
         public int api_created_ship_id { get; set; }
-        public int api_complete_time { get; set; }
+        public long api_complete_time { get; set; }
         public string api_complete_time_str { get; set; }
         public int api_item1 { get; set; }
         public int api_item2 { get; set; }
